Unity3D下增加元组、Lua方法、指针类型
Former-commit-id: f3a84b4d103d63fdec71af6b9deadc2287b840f5

+++ Source/Unity3D/UnityCommon/LuaUnityEnv.cpp
@@ -39,7 +39,7 @@ void LuaUnityEnv::setNativeObjectId(const void *instance, int nativeObjectId, st
 {
     if (_setNativeObjectIdHandler != NULL)
     {
-        _setNativeObjectIdHandler (instance, nativeObjectId, luaObjectId.c_str());
+        _setNativeObjectIdHandler ((long long)instance, nativeObjectId, luaObjectId.c_str());
     }
 }
